// src/pages/fatoresRisco/pressão Arterial/type.ts
export interface FormValues {
    nome?: string;
    idade?: string;
    etnia?: string;
    estatura?: string;
    peso?: string;
    fcRepouso?: string;
    nivelCondicionamento?: string;
    dataAfericao?: string;
    medicamento?: string;
    resultadoAfericao?: string;
    ingeriuEstimulantes?: string;
    estressado?: string;
    fumouUltimos30Min?: string;
    problemasSono?: string;
    comeuUltimos30Min?: string;
    atividadeFisicaUltimos30Min?: string;
    sistolica: string;
    diastolica: string;
  }